+++ src/Ignixa.FhirPath/Evaluation/Functions/StringFunctions.cs
@@ -1,0 +1,306 @@
+/*
+ * Copyright (c) 2025, Ignixa Contributors
+ *
+ * FhirPath string manipulation function implementations (Phase 23, Week 4).
+ * Implements indexOf(), substring(), startsWith(), endsWith(), upper(), lower(),
+ * length(), replace(), matches(), replaceMatches(), toChars(), join().
+ */
+
+using System.Text.RegularExpressions;
+using Ignixa.Abstractions;
+using Ignixa.FhirPath.Expressions;
+
+namespace Ignixa.FhirPath.Evaluation.Functions;
+
+/// <summary>
+/// String manipulation function implementations for FhirPath expressions.
+/// </summary>
+internal static class StringFunctions
+{
+    /// <summary>
+    /// indexOf() - Returns the 0-based index of the first occurrence of a substring.
+    /// Returns -1 if substring is not found.
+    /// </summary>
+    public static IEnumerable<ITypedElement> IndexOf(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count == 0)
+            throw new ArgumentException("indexOf() requires a substring argument");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var substringResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        if (substringResult?.Value is not string substring)
+            return Enumerable.Empty<ITypedElement>();
+
+        var index = str.IndexOf(substring, StringComparison.Ordinal);
+        return new[] { FunctionHelpers.CreateInteger(index) };
+    }
+
+    /// <summary>
+    /// substring() - Extracts a substring starting at a 0-based index.
+    /// Optionally accepts a length parameter.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Substring(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count == 0)
+            throw new ArgumentException("substring() requires a start argument");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var startResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        if (startResult?.Value is not int start)
+            return Enumerable.Empty<ITypedElement>();
+
+        if (start < 0 || start >= str.Length)
+            return Enumerable.Empty<ITypedElement>();
+
+        int? length = null;
+        if (arguments.Count > 1)
+        {
+            var lengthResult = evaluateExpression(focus, arguments[1], context).SingleOrDefault();
+            if (lengthResult?.Value is int len)
+                length = len;
+        }
+
+        var result = length.HasValue
+            ? str.Substring(start, Math.Min(length.Value, str.Length - start))
+            : str.Substring(start);
+
+        return new[] { FunctionHelpers.CreateString(result) };
+    }
+
+    /// <summary>
+    /// startsWith() - Tests if a string starts with a given prefix.
+    /// </summary>
+    public static IEnumerable<ITypedElement> StartsWith(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count == 0)
+            throw new ArgumentException("startsWith() requires a prefix argument");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var prefixResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        if (prefixResult?.Value is not string prefix)
+            return Enumerable.Empty<ITypedElement>();
+
+        return FunctionHelpers.ReturnBoolean(str.StartsWith(prefix, StringComparison.Ordinal));
+    }
+
+    /// <summary>
+    /// endsWith() - Tests if a string ends with a given suffix.
+    /// </summary>
+    public static IEnumerable<ITypedElement> EndsWith(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count == 0)
+            throw new ArgumentException("endsWith() requires a suffix argument");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var suffixResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        if (suffixResult?.Value is not string suffix)
+            return Enumerable.Empty<ITypedElement>();
+
+        return FunctionHelpers.ReturnBoolean(str.EndsWith(suffix, StringComparison.Ordinal));
+    }
+
+    /// <summary>
+    /// upper() - Converts a string to uppercase.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Upper(IEnumerable<ITypedElement> focus)
+    {
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        return new[] { FunctionHelpers.CreateString(str.ToUpperInvariant()) };
+    }
+
+    /// <summary>
+    /// lower() - Converts a string to lowercase.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Lower(IEnumerable<ITypedElement> focus)
+    {
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        // FhirPath lower() function explicitly requires lowercase, ToLowerInvariant is intentional
+#pragma warning disable CA1308 // Normalize strings to uppercase
+        return new[] { FunctionHelpers.CreateString(str.ToLowerInvariant()) };
+#pragma warning restore CA1308 // Normalize strings to uppercase
+    }
+
+    /// <summary>
+    /// length() - Returns the number of characters in a string.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Length(IEnumerable<ITypedElement> focus)
+    {
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        return new[] { FunctionHelpers.CreateInteger(str.Length) };
+    }
+
+    /// <summary>
+    /// replace() - Replaces all occurrences of a pattern string with a substitution string.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Replace(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count < 2)
+            throw new ArgumentException("replace() requires pattern and substitution arguments");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var patternResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        var substitutionResult = evaluateExpression(focus, arguments[1], context).SingleOrDefault();
+
+        if (patternResult?.Value is not string pattern || substitutionResult?.Value is not string substitution)
+            return Enumerable.Empty<ITypedElement>();
+
+        var result = str.Replace(pattern, substitution, StringComparison.Ordinal);
+        return new[] { FunctionHelpers.CreateString(result) };
+    }
+
+    /// <summary>
+    /// matches() - Tests if a string matches a regular expression pattern.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Matches(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count == 0)
+            throw new ArgumentException("matches() requires a regex argument");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var regexResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        if (regexResult?.Value is not string pattern)
+            return Enumerable.Empty<ITypedElement>();
+
+        try
+        {
+            var regex = new Regex(pattern);
+            return FunctionHelpers.ReturnBoolean(regex.IsMatch(str));
+        }
+        catch
+        {
+            return Enumerable.Empty<ITypedElement>();
+        }
+    }
+
+    /// <summary>
+    /// replaceMatches() - Replaces all regex pattern matches with a substitution string.
+    /// </summary>
+    public static IEnumerable<ITypedElement> ReplaceMatches(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        if (arguments.Count < 2)
+            throw new ArgumentException("replaceMatches() requires pattern and substitution arguments");
+
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        var patternResult = evaluateExpression(focus, arguments[0], context).SingleOrDefault();
+        var substitutionResult = evaluateExpression(focus, arguments[1], context).SingleOrDefault();
+
+        if (patternResult?.Value is not string pattern || substitutionResult?.Value is not string substitution)
+            return Enumerable.Empty<ITypedElement>();
+
+        try
+        {
+            var regex = new Regex(pattern);
+            var result = regex.Replace(str, substitution);
+            return new[] { FunctionHelpers.CreateString(result) };
+        }
+        catch
+        {
+            return Enumerable.Empty<ITypedElement>();
+        }
+    }
+
+    /// <summary>
+    /// toChars() - Splits a string into individual characters.
+    /// Returns a collection of single-character strings.
+    /// </summary>
+    public static IEnumerable<ITypedElement> ToChars(IEnumerable<ITypedElement> focus)
+    {
+        var list = focus.ToList();
+        if (list.Count != 1 || list[0].Value is not string str)
+            return Enumerable.Empty<ITypedElement>();
+
+        return str.Select(c => FunctionHelpers.CreateString(c.ToString()));
+    }
+
+    /// <summary>
+    /// join() - Concatenates a collection of strings with an optional separator.
+    /// If separator is not provided, concatenates without separator.
+    /// </summary>
+    public static IEnumerable<ITypedElement> Join(
+        IEnumerable<ITypedElement> focus,
+        IReadOnlyList<Expression> arguments,
+        EvaluationContext context,
+        Func<IEnumerable<ITypedElement>, Expression, EvaluationContext, IEnumerable<ITypedElement>> evaluateExpression)
+    {
+        var focusElements = focus.ToList();
+
+        // Get the separator (default to empty string if not provided)
+        var separator = string.Empty;
+        if (arguments.Count > 0)
+        {
+            var sepResult = evaluateExpression(focusElements, arguments[0], context).ToList();
+            if (sepResult.Count > 0 && sepResult[0].Value is string sepStr)
+            {
+                separator = sepStr;
+            }
+        }
+
+        // Concatenate all string values with the separator
+        var strings = focusElements
+            .Where(e => e.Value is string)
+            .Select(e => (string)e.Value!)
+            .ToList();
+
+        // join() always returns a string, even if empty
+        var result = string.Join(separator, strings);
+        return new[] { FunctionHelpers.CreateString(result) };
+    }
+}
